Protect user based actions
Golang api was unprotected. Use Auth0 JWT based authentication to send and verify
token on backend before allowing user to make modifications to protected resources.
Audience has to be a custom api identifier not the default auth0 one which is
the issuer for all tokens.
Update all frontend code to use bearer token